+++ easy.go
@@ -11,22 +11,22 @@ type EasyUserTokenPayload struct {
 	RestConfig
 	UserId string `json:"user_id"`
 	Email  string `json:"email"`
-	Name   string `json:"name"`
+	Name   string `json:"username"`
 	Gender int    `json:"gender"`
 	Birth  string `json:"birth"`
 	Phone  string `json:"phone"`
 }
 
-type EasyUserToken struct {
-	APIResponse
-	Data    struct {
-		UserToken        string `json:"user_token"`
-		ExpiredUnixtime  int64  `json:"expired_unixtime"`
-		ExpiredLocaltime string `json:"expired_localtime"`
-	} `json:"data"`
-}
+//type EasyUserToken struct {
+//	APIResponse
+//	Data    struct {
+//		UserToken        string `json:"user_token"`
+//		ExpiredUnixtime  int64  `json:"expired_unixtime"`
+//		ExpiredLocaltime string `json:"expired_localtime"`
+//	} `json:"data"`
+//}
 
-func (api *Api) GetUserToken(userToken EasyUserTokenPayload) (EasyUserToken, error) {
+func (api *Api) GetUserToken(userToken EasyUserTokenPayload) (APIResponse, error) {
 	if userToken.ApplicationId == "" {
 		userToken.ApplicationId = api.applicationId
 	}
@@ -38,14 +38,13 @@ func (api *Api) GetUserToken(userToken EasyUserTokenPayload) (EasyUserToken, err
 	req, err := api.NewRequest(http.MethodPost, "/request/user/token", body)
 	if err != nil {
 		errors.New("bootpay: ReserveCancelSubscribe error: " + err.Error())
-		return EasyUserToken{}, err
-	}
-	req.Header.Set("Authorization", api.token)
+		return APIResponse{}, err
+	} 
 	res, err := api.client.Do(req)
 
 	defer res.Body.Close()
 
-	result := EasyUserToken{}
+	result := APIResponse{}
 	json.NewDecoder(res.Body).Decode(&result)
 	return result, nil
 }